fix castform passive
